adjust dsgvo for report message
