+++ src/pages/Chat.tsx
@@ -11,10 +11,9 @@ import { useChat } from '../context/ChatContext';
 import { useLocation } from 'react-router-dom';
 import ChatContactsModal from '../modals/ChatContactsModal';
 import { useLayoutHelper } from '../hooks/useLayoutHelper';
-import ContactSupportModal from '../modals/ContactSupportModal';
+import ContactSupportModal, { ReportInfos } from '../modals/ContactSupportModal';
 import { MessageActionEvent } from 'talkjs/all';
 import { DateTime } from 'luxon';
-import { ReportInfos } from '../components/ContactSupportForm';
 
 const Chat: React.FC = () => {
     const inboxRef = useRef(null);
